+++ tsup.config.ts
@@ -3,6 +3,7 @@ import { defineConfig } from 'tsup';
 export default defineConfig({
     entry: {
         index: 'src/index.ts',
+        cli: 'src/cli.ts',
         puppeteer: 'src/integrations/puppeteer.ts',
         playwright: 'src/integrations/playwright.ts',
         'extower': 'src/integrations/extower.ts',